Fix PBR-01M

+++ packages/evm/contracts/Core.sol
@@ -27,7 +27,7 @@ abstract contract Core is Modifier {
 
     function _accruedAllowance(
         Allowance memory allowance,
-        uint256 blockTimestamp
+        uint64 blockTimestamp
     ) internal pure virtual returns (uint128 balance, uint64 timestamp);
 
     function _key(
